refactor(proxyMiddleware): create new file for each route handler
- Added a new file `countrySelect.ts` in the `handlers` directory to handle the GET and POST requests for country selection.
- The `get` function renders the `apply/which-country-list-do-you-want-to-be-added-to` view with the countries list, the previously selected answer, and a back link.
- The `post` function handles the form submission and validates the selected country. If the country is not valid, an error message is flashed and the user is redirected back to the country selection page. Otherwise, the selected country is stored in the session and the existence of the list for the selected country and service type is checked. If the list does not exist, the user is redirected to the "not-currently-accepting" page. Otherwise, the user is redirected to the next page based on the service type.
- Added `countrySelect` export in the `handlers/index.ts` file to export the `get` and `post` functions.
- Updated the `router.ts` file to import the `handlers` module using the `* as` syntax to include all exported functions from the `handlers` directory.
