+++ src/server/components/proxyMiddleware/handlers/countrySelect.ts
@@ -4,27 +4,7 @@ import { listExists, serviceTypeSchema } from "server/components/proxyMiddleware
 import { countriesList } from "server/services/metadata";
 import { camelCase } from "lodash";
 
-export const handlers = {
-  start: {
-    get: getStartPage,
-  },
-  countrySelect: {
-    get: getCountrySelectPage,
-    post: postCountrySelectPage,
-  },
-  stop: {
-    get: getStopPage,
-  },
-};
-
-function getStartPage(req: Request, res: Response) {
-  const { value: serviceType } = serviceTypeSchema.validate(req.params.serviceType);
-
-  req.session.application ??= {};
-  res.render(`apply/${serviceType}/start`);
-}
-
-function getCountrySelectPage(req: Request, res: Response) {
+export function get(req: Request, res: Response) {
   const { value: serviceType } = serviceTypeSchema.validate(req.params.serviceType);
 
   res.render("apply/which-country-list-do-you-want-to-be-added-to", {
@@ -34,7 +14,7 @@ function getCountrySelectPage(req: Request, res: Response) {
   });
 }
 
-async function postCountrySelectPage(req: Request, res: Response) {
+export async function post(req: Request, res: Response) {
   const { country } = req.body;
   const validatedCountry = validateCountryLower(country);
   const { value: serviceType } = serviceTypeSchema.validate(req.params.serviceType);
@@ -63,19 +43,4 @@ async function postCountrySelectPage(req: Request, res: Response) {
   }
 
   res.redirect(`/application/${serviceType}/${nextPagePath[serviceType!]}`);
-}
-
-function getStopPage(req: Request, res: Response) {
-  const { value: serviceType } = serviceTypeSchema.validate(req.params.serviceType);
-  const serviceTitles = {
-    lawyers: "lawyers",
-    "funeral-directors": "funeral directors",
-    "translators-interpreters": "translators and interpreters",
-  };
-
-  res.render("apply/not-accepting-currently", {
-    backLink: `/application/${serviceType}/which-country-list-do-you-want-to-be-added-to`,
-    country: req.session?.application?.country,
-    serviceTitle: serviceTitles[serviceType as keyof typeof serviceTitles],
-  });
 }
